일부 코드 정리
- Long -> Int
- 화면 전환 코드 변경
- MappingDao Warnings 처리
- 미사용 코드 삭제

+++ app/src/main/java/com/blueland/kimonesheet/db/entity/FolderEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "Folder")
 data class FolderEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long? = null,
+    @PrimaryKey(autoGenerate = true) val id: Int? = null,
     @ColumnInfo val name: String,
     @ColumnInfo(name = "reg_date") val regDate: Long = System.currentTimeMillis(),
     @ColumnInfo(name = "mod_date") val modDate: Long = System.currentTimeMillis(),
