+++ app/src/main/java/com/blueland/kimonesheet/db/entity/MappingEntity.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "Mapping")
 data class MappingEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long,
-    @ColumnInfo(name = "parent_id") val parentId: Long,
-    @ColumnInfo(name = "child_id") val child_id: Long,
-    @ColumnInfo(defaultValue = "0") var type: Long
+    @PrimaryKey(autoGenerate = true) val id: Int,
+    @ColumnInfo(name = "parent_id") val parentId: Int,
+    @ColumnInfo(name = "child_id") val child_id: Int,
+    @ColumnInfo(defaultValue = "0") var type: Int
 )
