+++ app/src/main/java/com/blueland/kimonesheet/db/entity/MemoEntity.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
 
 @Entity(tableName = "Memo")
 data class MemoEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long? = null,
+    @PrimaryKey(autoGenerate = true) val id: Int? = null,
     @ColumnInfo val title: String,
     @ColumnInfo val content: String,
     @ColumnInfo(name = "reg_date") val regDate: Long = System.currentTimeMillis(),
